adds updated condition for while

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -12,8 +12,8 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int count = 0;
-	unsigned int x;
+	unsigned long int count = 0;
+	unsigned long int x;
 
 	x = n ^ m;
